feat(picker): convert to typescript
