+++ src/components/Picker/PickerComponent.android.tsx
@@ -1,37 +1,18 @@
 import * as React from "react";
-import {
-  View,
-  Picker as NativePicker,
-  StyleSheet,
-  StyleProp,
-  ViewStyle,
-} from "react-native";
+import { View, Picker as NativePicker, StyleSheet } from "react-native";
 import { withTheme } from "../../core/theming";
 
-import TextField, { Props as TextFieldProps } from "../TextField";
+import TextField from "../TextField";
 import Touchable from "../Touchable";
+import { PickerComponentProps } from "./PickerTypes";
 
-interface PickerOption {
-  value: string;
-  label: string;
-}
-
-interface Props extends TextFieldProps {
-  style?: StyleProp<ViewStyle>;
-  options: PickerOption[];
-  placeholder?: string;
-  selectedValue: string;
-  disabled?: boolean;
-  onValueChanged: (value: string, index: number) => void;
-}
-
-const Picker: React.FC<Props> = ({
+const Picker: React.FC<PickerComponentProps> = ({
   style,
   options,
   placeholder,
   selectedValue,
   disabled = false,
-  onValueChanged: onValueChangeOverride = () => {},
+  onValueChange: onValueChangeOverride = () => {},
   ...props
 }) => {
   const textField = React.useRef<typeof TextField | undefined>(undefined);
